Add AC device controller to ACState task add dialogue #9
